padColumns: switch to using string option instead of bool
With bool, the way the logic is constructed means there's not an easy
way to have the option default to true. Would need to check if the flag
was set in the CLI, then if not (and if plugin-table was enabled), set
the value to true.

It's just way easier to do a string and default to "on", so that's how
we're doing things!